+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* added the optional `sessionPoolMaxSize` client setting to limit the in-memory session pool size
 
 ## 1.16.3
 * improve log